requirements.txt
# Strategy Ensemble - Requirements
# ================================

# Core data processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0

# Database
duckdb>=0.9.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0
optuna>=3.6.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# HTTP requests
requests>=2.31.0
httpx>=0.24.0

# LLM API
google-genai>=1.0.0

# Scheduling
schedule>=1.2.0
apscheduler>=3.10.0

# Logging
loguru>=0.7.0

# Validation
pydantic>=2.0.0

# Telegram
telethon>=1.34.0

# Visualization (optional but recommended)
matplotlib>=3.7.0
seaborn>=0.12.0

# Development (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
ruff>=0.1.0
mypy>=1.5.0

# WebSocket (KIS real-time data)
websockets>=12.0
antrophic
ccxt
# Jupyter (optional)
# jupyterlab>=4.0.0
# ipykernel>=6.25.0
